Allow sudo in Travis CI
Allows the use of sudo in Travis CI in an attempt to fix the test suite.
Allowing sudo causes Travis to execute the build in a different
environment.

As part of this commit, the "skip tests by default" functionality from
pull request #21 is changed to "run tests by default". The addition of
the flag `-DskipTests` skips any tests as before.