Created basic web service
Created a basic Calamari web service. The service, which can be run via
`./gradlew web:run`, has one endpoint, `POST /prawn`. The endpoint
accepts `useSBM` and `userLinFits` as optional URL parameters, but
requires a multipart form with a `prawnFile` file provided as the
payload. It then responds with a ZIP of reports if the request is
successful.